requirements.txt
django~=4.2.0
django-environ==0.11.2
djangorestframework==3.15.2
markdown==3.6
django-filter==24.2
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
model-bakery==1.18.2